Decouple registry auth adapters for Base and Para
