+++ apps/registry/src/components/control-bar/network-select.tsx
@@ -3,6 +3,7 @@
 import { useState, type FC } from "react";
 import { ChevronDownIcon, CheckIcon } from "lucide-react";
 import { cn, SUPPORTED_CHAINS, getChainInfo } from "@aomi-labs/react";
+import type { Chain } from "viem";
 import { Button } from "@/components/ui/button";
 import {
   Popover,
@@ -14,17 +15,18 @@ import { useAomiAuthAdapter } from "../../lib/aomi-auth-adapter";
 export type NetworkSelectProps = {
   className?: string;
   /** Override the default chain list from the lib */
-  chains?: typeof SUPPORTED_CHAINS;
+  chains?: readonly Chain[];
 };
 
 export const NetworkSelect: FC<NetworkSelectProps> = ({
   className,
-  chains = SUPPORTED_CHAINS,
+  chains,
 }) => {
   const adapter = useAomiAuthAdapter();
   const { chainId, isConnected } = adapter.identity;
   const switchChain = adapter.switchChain;
   const isPending = adapter.isSwitchingChain;
+  const selectableChains = chains ?? adapter.supportedChains ?? SUPPORTED_CHAINS;
   const [open, setOpen] = useState(false);
 
   // Show only when wallet is connected.
@@ -58,7 +60,7 @@ export const NetworkSelect: FC<NetworkSelectProps> = ({
         className="w-[180px] rounded-3xl p-1 shadow-none"
       >
         <div className="flex flex-col gap-0.5">
-          {chains.map((chain) => (
+          {selectableChains.map((chain) => (
             <button
               key={chain.id}
               disabled={isPending || !switchChain}
