+++ apps/registry/src/index.ts
@@ -25,14 +25,19 @@ export {
 export type { UserConfig } from "@aomi-labs/react";
 export { formatAddress, getNetworkName } from "@aomi-labs/react";
 export type { AomiAuthAdapter } from "./lib/aomi-auth-adapter";
-export { useAomiAuthAdapter } from "./lib/aomi-auth-adapter";
+export {
+  AomiAuthAdapterProvider,
+  AomiAuthRuntimeUserSync,
+  useAomiAuthAdapter,
+} from "./lib/aomi-auth-adapter";
+export { AomiBaseAccountProvider } from "./lib/aomi-auth-adapter/providers/base-account";
 export type {
   AomiAuthIdentity,
   AomiAuthStatus,
-} from "./lib/auth-identity";
+} from "./lib/aomi-auth-adapter";
 export {
   AOMI_AUTH_BOOTING_IDENTITY,
   AOMI_AUTH_DISCONNECTED_IDENTITY,
   formatAuthProvider,
   inferAuthProvider,
-} from "./lib/auth-identity";
+} from "./lib/aomi-auth-adapter";
