+++ apps/registry/src/lib/aomi-auth-adapter/index.ts
@@ -1,0 +1,6 @@
+"use client";
+
+export * from "./context";
+export * from "./identity";
+export * from "./runtime-user-sync";
+export * from "./types";
